distro/generic: move TestBootstrapContainers() to external tests
Move TestISOLabel() to a new test file called 'distro_internal_test'.
The 'distro_test' file now holds external tests (package generic_test).
